expand ausdauer api, add more data
