+++ app/api/v1/ausdauer/index.js
@@ -2,11 +2,11 @@ module.exports = (express, db) => {
   // get routes from express
   const ausdauer = express.Router();
   // get sql connection
-  const tblAusdauer = require("../../../logic/ausdauer/queries")(db);
-
+  const tblAusdauer = require("../../../logic/queries")(db, "ausdauer");
   // routes
-  ausdauer.get("/:gender", tblAusdauer.getAll);
-
+  ausdauer.get("/", tblAusdauer.getAll);
+  ausdauer.get("/:gender", tblAusdauer.getAllFromGender);
+  ausdauer.get("/:gender/:age", tblAusdauer.getAllFromGenderAndAge);
   // return api
   return ausdauer;
 }
