+++ sql/01-create-db.sql
@@ -1,0 +1,2 @@
+DROP DATABASE IF EXISTS sportabzeichen;
+CREATE DATABASE sportabzeichen;
