+++ sql/02-create-and-fill-table-ausdauer.sql
@@ -1,0 +1,50 @@
+\c sportabzeichen;
+
+CREATE TABLE ausdauer (
+  ID SERIAL PRIMARY KEY,
+  geschlecht VARCHAR,
+  jahr INTEGER,
+  uebung VARCHAR,
+  alt_von INTEGER,
+  alt_bis INTEGER,
+  abzeichen_gold VARCHAR,
+  abzeichen_silber VARCHAR,
+  abzeichen_bronze VARCHAR,
+  einheit VARCHAR
+);
+
+INSERT INTO ausdauer (geschlecht, jahr, uebung, alt_von, alt_bis, abzeichen_gold, abzeichen_silber, abzeichen_bronze, einheit) VALUES 
+  ('m', 2016, '3.000 m Lauf', 18, 19, '13:20', '15:20', '17:20', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 20, 24, '12:50', '14:50', '16:50', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 25, 29, '13:10', '15:10', '17:10', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 30, 34, '14:10', '16:10', '18:40', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 35, 39, '15:00', '17:20', '19:50', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 40, 44, '15:50', '18:30', '21:00', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 45, 49, '16:30', '19:30', '22:10', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 50, 54, '17:20', '20:20', '23:20', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 55, 59, '17:50', '20:50', '23:50', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 60, 64, '18:30', '21:30', '24:30', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 65, 69, '19:00', '22:00', '25:00', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 70, 74, '19:20', '22:20', '25:20', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 75, 79, '20:00', '23:00', '26:00', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 80, 84, '20:30', '23:30', '26:30', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 85, 89, '21:30', '24:30', '27:30', 'Min'),
+  ('m', 2016, '3.000 m Lauf', 90, 99, '23:50', '26:50', '29:50', 'Min');
+
+INSERT INTO ausdauer (geschlecht, jahr, uebung, alt_von, alt_bis, abzeichen_gold, abzeichen_silber, abzeichen_bronze, einheit) VALUES
+  ('f', 2016, '3.000 m Lauf', 18, 19, '18:00', '20:00', '22:00', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 20, 24, '17:50', '19:50', '21:50', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 25, 29, '18:00', '19:50', '22:00', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 30, 34, '18:10', '20:10', '22:20', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 35, 39, '18:20', '20:30', '23:10', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 40, 44, '18:30', '21:00', '24:00', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 45, 49, '18:50', '21:30', '24:40', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 50, 54, '19:10', '22:10', '25:10', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 55, 59, '19:50', '22:50', '25:50', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 60, 64, '20:30', '23:30', '26:30', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 65, 69, '21:10', '24:10', '27:10', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 70, 74, '21:40', '24:40', '27:40', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 75, 79, '22:30', '25:20', '28:30', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 80, 84, '23:40', '26:40', '29:40', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 85, 89, '25:00', '27:50', '31:00', 'Min'),
+  ('f', 2016, '3.000 m Lauf', 90, 99, '26:30', '29:30', '32:30', 'Min');
